# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PerioVision AI API",
    "version": "2.0",
    "description": "Research prototype for periodontal decision support. Every JSON response uses the envelope {data, meta, error, mode}. Authenticate with POST /api/auth/login, then send 'Authorization: Bearer <access_token>'. x-permission names the RBAC permission each operation requires (see docs/SECURITY.md)."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Envelope": {
        "type": "object",
        "required": [
          "data",
          "meta",
          "error",
          "mode"
        ],
        "properties": {
          "data": {
            "description": "Payload (object, list or null on error)"
          },
          "meta": {
            "type": "object",
            "description": "Extra information such as counts"
          },
          "error": {
            "nullable": true,
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          },
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ]
          }
        }
      },
      "LoginIn": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "MfaIn": {
        "additionalProperties": false,
        "properties": {
          "mfa_token": {
            "maxLength": 4096,
            "minLength": 10,
            "title": "Mfa Token",
            "type": "string"
          },
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "mfa_token",
          "code"
        ],
        "title": "MfaIn",
        "type": "object"
      },
...
      "TotpCodeIn": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "TotpCodeIn",
        "type": "object"
      },
...
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Sign in with email and password (returns an access token or an MFA challenge)",
        "operationId": "auth_login_post",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Sign out and revoke this session",
        "operationId": "auth_logout_post",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "self:manage"
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Current user and their permissions",
        "operationId": "auth_me_get",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "self:manage"
      }
    },
    "/api/auth/mfa": {
      "post": {
        "summary": "Complete sign-in with a 6-digit TOTP code",
        "operationId": "auth_mfa_step_post",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaIn"
              }
            }
          }
        }
      }
    },
    "/api/auth/mfa/confirm": {
      "post": {
        "summary": "Confirm TOTP enrolment with a code",
        "operationId": "auth_mfa_confirm_post",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "self:manage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpCodeIn"
              }
            }
          }
        }
      }
    },
    "/api/auth/mfa/disable": {
      "post": {
        "summary": "Turn TOTP off (needs a current code)",
        "operationId": "auth_mfa_disable_post",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          },
          "422": {
            "description": "Validation failed; error.details lists each field"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "self:manage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpCodeIn"
              }
            }
          }
        }
      }
    },
    "/api/auth/mfa/enroll": {
      "post": {
        "summary": "Start TOTP enrolment (QR code)",
        "operationId": "auth_mfa_enroll_post",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "self:manage"
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Get a new access token from the refresh cookie (rotates it)",
        "operationId": "auth_refresh_post",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/auth/sessions": {
      "get": {
        "summary": "My active sessions",
        "operationId": "auth_my_sessions_get",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid token / session (Zero Trust check)"
          },
          "403": {
            "description": "Role lacks the permission, or account blocked"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "self:manage"
      }
    },
...
  }
}